+++ source/FileHandler.hpp
@@ -19,12 +19,12 @@ class NotEnoughData : public std::exception {
 
 class FileHandler {
     protected:
-        std::queue<int>* pbuff;
+        std::queue<uint8_t>* pbuff;
         std::fstream file;
         char* charBuff;
         int bufferSize;
     public:
-        FileHandler(std::queue<int>* pbuff, int bufferSize) 
+        FileHandler(std::queue<uint8_t>* pbuff, int bufferSize) 
         : pbuff(pbuff), bufferSize(bufferSize) {
             charBuff = new char [bufferSize];
         }
